Create Kumade::Packager. More one step to close issue #7.
